+++ src/Sentry/GlobalSessionManager.cs
@@ -253,7 +253,7 @@ internal class GlobalSessionManager : ISessionManager
         if (status == SessionEndStatus.Exited && session.IsMarkedAsPendingUnhandled)
         {
             status = SessionEndStatus.Unhandled;
-            _options.LogDebug("Session ended as 'Unhandled' due to pending status.");
+            _options.LogDebug("Session ended with pending 'Unhandled' (but not `Terminal`) exception.");
         }
 
         if (status == SessionEndStatus.Crashed)
@@ -299,7 +299,7 @@ internal class GlobalSessionManager : ISessionManager
 
         var now = _clock.GetUtcNow();
         _lastPauseTimestamp = now;
-        PersistSession(session.CreateUpdate(false, now), now);
+        PersistSession(session.CreateUpdate(false, now), now, session.IsMarkedAsPendingUnhandled);
     }
 
     public IReadOnlyList<SessionUpdate> ResumeSession()